Changed the data type of position in hash_table_get_element to size_t

+++ hashtable.c
@@ -91,7 +91,7 @@ int hash_table_has_element(hash_table_t *hash_table, char *string) {
 }
 
 hash_table_list_t *hash_table_get_element(hash_table_t *hash_table, char *string) {
-    unsigned int position = hash_table_get_position(hash_table, string);
+    size_t position = hash_table_get_position(hash_table, string);
     hash_table_list_t *list = hash_table->elements[position];
 
 	//Same as hash_table_has_element, loop through all the entries in the cell
